chore: split files into backend and factory
