+++ vault-plugins/flant_iam/backend/backend.go
@@ -3,9 +3,7 @@ package backend
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"net/http"
-	"strings"
 
 	"github.com/hashicorp/errwrap"
 	"github.com/hashicorp/go-uuid"
@@ -13,50 +11,6 @@ import (
 	"github.com/hashicorp/vault/sdk/helper/jsonutil"
 	"github.com/hashicorp/vault/sdk/logical"
 )
-
-var _ logical.Factory = Factory
-
-// Factory configures and returns Mock backends
-func Factory(ctx context.Context, conf *logical.BackendConfig) (logical.Backend, error) {
-	if conf == nil {
-		return nil, fmt.Errorf("configuration passed into backend is nil")
-	}
-
-	b := newBackend()
-
-	if err := b.Setup(ctx, conf); err != nil {
-		return nil, err
-	}
-
-	return b, nil
-}
-
-func newBackend() logical.Backend {
-	b := &framework.Backend{
-		Help:        strings.TrimSpace(commonHelp),
-		BackendType: logical.TypeLogical,
-	}
-
-	backendLayer := &layerBackend{b}
-
-	tenantKeys := &keyManager{
-		idField:   "tenant_id",
-		entryName: "tenant",
-	}
-
-	userKeys := &keyManager{
-		idField:   "user_id",
-		entryName: "user",
-		parent:    tenantKeys,
-	}
-
-	b.Paths = framework.PathAppend(
-		backendLayer.paths(tenantKeys),
-		backendLayer.paths(userKeys),
-	)
-
-	return b
-}
 
 type layerBackend struct {
 	logical.Backend
@@ -258,7 +212,3 @@ func errNotFoundResponse(req *logical.Request, key string) *logical.Response {
 	resp, _ := logical.RespondWithStatusCode(errResp, req, http.StatusNotFound)
 	return resp
 }
-
-const commonHelp = `
-IAM API here
-`
